Move CUDA code in the cuda files and clean the rest.  Rename files to reflect that they don't need nvcc anymore.
